Pull out common purchaseHistory setup, move reusable tests to main
