+++ test-utils/src/bc5SupportDev/java/com/revenuecat/purchases/utils/stubsBillingClient5.kt
@@ -1,9 +1,16 @@
 package com.revenuecat.purchases.utils
 
 import com.android.billingclient.api.BillingClient
+import com.android.billingclient.api.BillingResult
 import com.android.billingclient.api.Purchase
 import com.android.billingclient.api.PurchaseHistoryRecord
+import com.android.billingclient.api.PurchaseHistoryResponseListener
+import com.android.billingclient.api.QueryPurchaseHistoryParams
 import com.android.billingclient.api.SkuDetails
+import io.mockk.every
+import io.mockk.mockk
+import io.mockk.mockkStatic
+import io.mockk.slot
 import org.json.JSONArray
 
 @SuppressWarnings("LongParameterList", "MagicNumber")
@@ -71,3 +78,39 @@ fun stubPurchaseHistoryRecord(
                 "purchaseToken": "$purchaseToken"
             }
         """.trimIndent(), signature)
+
+fun BillingClient.mockQueryPurchaseHistory(
+    result: BillingResult,
+    history: List<PurchaseHistoryRecord>
+): QueryPurchaseHistoryParams.Builder {
+    mockkStatic(QueryPurchaseHistoryParams::class)
+
+    val mockBuilder = mockk<QueryPurchaseHistoryParams.Builder>(relaxed = true)
+    every {
+        QueryPurchaseHistoryParams.newBuilder()
+    } returns mockBuilder
+
+    every {
+        mockBuilder.setProductType(any())
+    } returns mockBuilder
+
+    val params = mockk<QueryPurchaseHistoryParams>(relaxed = true)
+    every {
+        mockBuilder.build()
+    } returns params
+
+    val billingClientPurchaseHistoryListenerSlot = slot<PurchaseHistoryResponseListener>()
+
+    every {
+        queryPurchaseHistoryAsync(
+            params,
+            capture(billingClientPurchaseHistoryListenerSlot)
+        )
+    } answers {
+        billingClientPurchaseHistoryListenerSlot.captured.onPurchaseHistoryResponse(
+            result,
+            history
+        )
+    }
+    return mockBuilder
+}
